[WIP][FEAT] input_choice: modular options
This makes input_choice more modular an thus allowing nested options or
numbers as options.

It removes also the usage of 'return_hotkey' as the option itself can
now determine what is returned and the index doesn't make sense for
non-standard options. It emits a warning if it's set to True (by default
it was False anyway) and fails if the original answers array didn't just
consist of tuples.

Todo:
- Validate uniquity
- Make option somehow immutable

Change-Id: Ib922bcdb77ee34e8d0268ded2b003e7e6a9005b4