[FEAT] Get shared FilePage and make checks dynamic
With version 1.22 the meta property 'filerepoinfo' was introduced which
allows pywikibot to dynamically detect if a file is remote and also to
create a FilePage instance from another FilePage instance which is in
the remote repository (to do checks there for example).

To support older wikis, families can use 'get_file_repositories' which
returns a dictionary. It contains for each repository name either False
(if it's local) or a URL which can be parsed by pywikibot.Site() (if
it's remote).

Because in not all cases 'filerepoinfo' returns unique features it's not
always possible to determine the site via a name (see T85153), so even
with MW 1.22 there might be remote repositories needed to be defined via
'get_file_repositories'.

This changes the behavior of imageuncat: It now uses directly the site
and not the imagerepository of it. This allows the usage of the script
(in theory) on any site and not just on those which are defined via
imagerepository.

It also introduces a backwards compatible mode if neither
get_file_repositories nor MW 1.22 is used. It guesses the repository
names and uses 'shared_image_repository'. To avoid guessing the names it
is possible to define them explicit by adding a 'image_repository_names'
method to the family which returns the local repository name and remote
repository name.

Bug: T74847
Change-Id: Iab8b8dbbccfe5294223096586a00bb88d94a11ac