[FEAT] Throttle: SQLite to avoid race conditions
This largly restructures how Throttle works. Instead of slowly
increasing delays it is now enqueing itself and the next request to that
site has to wait for it.

It also uses a SQLite database and avoids with that race conditions as
other SQLite modifications need to wait for previous ones which can
happily increase the given time.

Also most of the attributes have been renamed or hidden (via an
underscore). The maxthrottle configuration value has been deprecated.

Change-Id: I800b2f1e84c0ee3f57bda79ddce46d216c698e83