[IMPROV] Site: Return preloads in requested order
Instead of returning the preloaded page in the order returned by the API
it caches the pages and returns them in the same order as requested.

Removed the tests using unexpected titles as they are now broken. Page.title()
always returns a normalized title so manipulating that in the test cannot be
anticipated by the library.

Change-Id: Ia554e81881e12e0c4d00f45d19b484baccd5e447