[FIX] Use str for __repr__ in both Python versions
This returns str in both Python versions (which essentially means bytes
in Python 2 and unicode in Python 3). Because of encoding problems it
forces some unicode strings back into bytes with Python 2. As repr()
should always return str it needs to be decoded when it's added to a
unicode. And with 1e54a7d6 a lot of those str have been converted into
unicode so it'll now decode it (while before it inserting a bytes into
bytes) using ASCII which is not always possible. So it's forcing strings
which have been bytes before 1e54a7d6 into bytes again by using 'str()'.

This only applies to Python 2 and it will only change some parts in
Python 3 whenever repr() returned bytes instead of str.

Bug: T95499
Change-Id: Ic7c96194c0d975033923ea43af19f9f40d331321