[FIX] Allow apostrophe in password
If the password contains an apostrophe it'll mess up the line in the password
file. E.g. if the password is "foo'bar" it'll add the following line:

  ('username', 'foo'bar')

This escapes now any quotes and it'll generate the following line:

  ('username', 'foo\'bar')

Bug: T105783
Change-Id: I45501958362ef07e20b73448a11f11a4160f3136
(cherry picked from commit ac07c7a09ca43c66bb5352533ac88d766d64bdbf)