[IMPROV] script_test: Simplify test for stderr
If there is expected output for that script it should just check that. The
additional check if it's an autorun script are unnecessary. The same is true
for the reverse and if there is no expected output and it is not an autorun
script then it should do the checks related to that.

Change-Id: Ic40aee0b1975ad1f785f27bafd3e56bc09511ada
(cherry picked from commit 00eb20fb0b3c1e801cdc08a12f477472600dcb73)