[FIX] QueryGenerator: buffer data before yielding
QueryGenerator yields all items in resultdata, but there is no
guarantee that all data for an item arrived in one response.
Some data will appear in the following response (see cc6d7b4).

This patch buffers results until all data are for an item are fetched.
This is based on the fact that the API, when query-continuing, keeps on
repeating the same pages until all requested data are fetched.

Skips two langlink tests like in deb64e80 because the interwiki map of the beta
wiki is to production wikis.

Change-Id: Iccb3a96b0248fdab0650edfda23d05ecec0dadbd