[FEAT] APISite: Generic language prefix list
Instead of relying on Family having defined all languages it can return all
prefixes used as langlinks by using the interwiki map.

This is changing how the `_InterwikiMap` class works as it's now possible to
iterate over it and handle it like a normal mapping without having to worry
about it creating site instances. Instead `get_checked()` now checks if the
site instance is not actually an exception.

It also set the interwiki map of `DrySite` to an empty dict to avoid queries
for the interwiki entries. With `add_IWEntry` it's easily possible to add a new
entry to that mapping.

Change-Id: Ib2c080d34a5ac70a5e6c09da21cf7dbcb27df2f8