[FEAT] Make FrozenDict like MappingProxyType
In Python 2 there is not `MappingProxyType` and even the `MappingProxyType` in
Python 3 is not subclassable while it might be advantageous to have a class
holding a dict which is subclassable.

This also provides `MappingProxyType` in Python 2 like in Python 3 with the
difference that it's subclassable and has the `create` classmethod.

Change-Id: I25845b239c7d200168679b7cc737257502985940