[IMPROV] Don't limit cryptography module on 2.6
The cryptography module does complain on Python 2.6 that it's outdated. So it's
using the newest version which doesn't issue that warning. But that could
compromise the security as it won't include security patches.

This is now not limiting that version anymore and installing the newest version
on Python 2.6 as well. By default the deprecation warning is hidden but if the
user still wishes to suppress this warning they can install version 0.8.2 or
before.

Change-Id: If5270a0b2214c2f3b124d50180a122bfeeadc3cb