[IMPROV] Enable 3.6 builds
After the 3.6 build has been disabled with a3b175f0 because pywikibot used
`getargspec` which was removed in 3.6 and because of Python's bug 24667 [1].
Now pywikibot itself doesn't use `getargspec` with baf31f1b and Python's bug
has been fixed so it's again possible to run Python 3.6 builds.

Because Python 3.6 removed `getargspec` it also removed the `ArgSpec` class
which requires the `tests.tools_tests.TestPythonArgSpec` to implement a
property to return the expected class. It should be skipped on 3.6 anyway but
the class was of course still parsed.

The additional packages installed for `requests` are not installed anymore on
Python 3, as it should be only necessary on Python 2.6 and Python <2.7.9. And
the `cryptography` package still uses `getargspec`.

[1]: http://bugs.python.org/issue24667

Change-Id: I5f84d51c1b4a75cc875df2d27ac14042d6119232