[FIX] aspects: Imply user on write tests
At the moment all writes require a user to be logged in but tests may not
define user and in that case the write test might fail because it's not working
on a logged in user. This sets user to True if it hasn't been set and is a
write test. So in case someone writes a write test without requiring a logged
in user it's necessary to explicitly set write to False now.

Change-Id: I4ef37ed6d9b28b72c0a1aa6be447818408c7107f
(cherry picked from commit 6255530f833c536a60d48197b54af9c5206405c1)