[IMPROV] Site: Return preloads in requested order
Instead of returning the preloaded page in the order returned by the API
it caches the pages and returns them in the same order as requested.

It also needs to modify the tests using unexpected titles because
preloadpages() now assumes that the Page title are sanitized and changing the
link's title after it has been parsed can't be anticipated. Instead it's
reparsing the link.

Change-Id: Ia554e81881e12e0c4d00f45d19b484baccd5e447