[IMPROV] i18n: Deprecate twntranslate
This deprecates `twntranslate` in favor of `twtranslate`. It also modifies
`_extract_plural` to require a mapping. `twntranslate` creates a dummy mapping
to simulate the deprecated types like lists to be used by `_extract_plural`.

This dummy mapping is also used by `translate` in order to deprecate the
non-mapping usage of parameters there. Not using a mapping for parameters in
i18n is only sensible if there is just one value but otherwise it might be that
parameters change the order in which case a list or tuple will produce
incorrect results. For the sake of consistency this deprecates all usage of
non-mapping parameters.

Change-Id: I5b97b10fbe0bab53993781c983db92c0aec30b30